chore: move chunkArray function to utilities
